Update Search_species.R for new rfishbase version
Been running into a few issues maintaining the older version of rfishbase that had the taxa table built in. This solution works on my end with the newer versions of rfishbase (Currently running 4.0.0). This version loads the taxa table once the first time it is called and prints out some annoying text, but caches the resulting taxa table for future calls within a session. One downside is that means an internet connection is required. Thoughts? Would need to update DESCRIPTION file as well to something like     rfishbase (>= 4.0.0) or whenever that change was made